LPS-65031 Move common logic to a base class and extend from there
