+++ modules/apps/foundation/users-admin/users-admin-demo-data-creator-impl/src/main/java/com/liferay/users/admin/demo/data/creator/internal/SiteAdminUserDemoDataCreatorImpl.java
@@ -19,20 +19,9 @@ import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.model.RoleConstants;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.model.UserConstants;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.users.admin.demo.data.creator.SiteAdminUserDemoDataCreator;
-
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.List;
-import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -42,80 +31,23 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(service = SiteAdminUserDemoDataCreator.class)
 public class SiteAdminUserDemoDataCreatorImpl
-	implements SiteAdminUserDemoDataCreator {
+	extends BaseUserDemoDataCreator implements SiteAdminUserDemoDataCreator {
 
 	public User create(long groupId, String emailAddress)
 		throws PortalException {
 
 		Group group = _groupLocalService.getGroup(groupId);
 
-		User user = _userLocalService.fetchUserByEmailAddress(
-			group.getCompanyId(), emailAddress);
+		User user = createBaseUser(group.getCompanyId(), emailAddress);
 
-		if (user != null) {
-			return user;
-		}
-
-		String emailAccountName = emailAddress.substring(
-			0, emailAddress.indexOf(StringPool.AT));
-
-		String[] fullNameArray = StringUtil.split(
-			emailAccountName, StringPool.PERIOD);
-
-		String firstName = StringUtil.randomString();
-		String lastName = StringUtil.randomString();
-
-		if (fullNameArray.length > 0) {
-			firstName = StringUtil.upperCaseFirstLetter(fullNameArray[0]);
-		}
-
-		if (fullNameArray.length > 1) {
-			lastName = StringUtil.upperCaseFirstLetter(fullNameArray[1]);
-		}
-
-		boolean autoPassword = true;
-		String password1 = "test";
-		String password2 = "test";
-		long facebookId = 0;
-		String openId = StringPool.BLANK;
-		Locale locale = LocaleUtil.SPAIN;
-		String middleName = StringPool.BLANK;
-		long prefixId = 0;
-		long suffixId = 0;
-		boolean male = true;
-		int birthdayMonth = Calendar.OCTOBER;
-		int birthdayDay = 18;
-		int birthdayYear = 1985;
-		String jobTitle = "Test Engineer";
-		long[] organizationIds = null;
-		long[] roleIds = null;
-		long[] userGroupIds = null;
-		boolean sendMail = false;
-
-		user = _userLocalService.addUser(
-			UserConstants.USER_ID_DEFAULT, group.getCompanyId(), autoPassword,
-			password1, password2, true, StringPool.BLANK, emailAddress,
-			facebookId, openId, locale, firstName, middleName, lastName,
-			prefixId, suffixId, male, birthdayMonth, birthdayDay, birthdayYear,
-			jobTitle, new long[] {groupId}, organizationIds, roleIds,
-			userGroupIds, sendMail, new ServiceContext());
+		userLocalService.setGroupUsers(groupId, new long[] {user.getUserId()});
 
 		Role role = _roleLocalService.getRole(
 			group.getCompanyId(), RoleConstants.SITE_ADMINISTRATOR);
 
-		_userLocalService.addRoleUser(role.getRoleId(), user);
+		userLocalService.addRoleUser(role.getRoleId(), user);
 
-		user = _userLocalService.getUser(user.getUserId());
-
-		_users.add(user);
-
-		return user;
-	}
-
-	public void delete() throws PortalException {
-		for (User user : _users) {
-			_userLocalService.deleteUser(user);
-		}
+		return userLocalService.getUser(user.getUserId());
 	}
 
 	@Reference(unbind = "-")
@@ -128,14 +60,7 @@ public class SiteAdminUserDemoDataCreatorImpl
 		_roleLocalService = roleLocalService;
 	}
 
-	@Reference(unbind = "-")
-	protected void setUserLocalService(UserLocalService userLocalService) {
-		_userLocalService = userLocalService;
-	}
-
 	private GroupLocalService _groupLocalService;
 	private RoleLocalService _roleLocalService;
-	private UserLocalService _userLocalService;
-	private final List<User> _users = new ArrayList();
 
 }
